IMPORTANT CHANGES
Reworked data synchronization (IMPORTANT)
Reworked authentication logic (IMPORTANT)
Fixes in save training and challenges results to DB (IMPORTANT)

+++ app/src/main/java/com/myrungo/rungo/profile/stats/models/Challenge.kt
@@ -3,7 +3,7 @@ package com.myrungo.rungo.profile.stats.models
 data class Challenge(
     val distance: Double = 0.0,
 
-    val hour: Int = 0,
+    val hour: Double = 0.0,
 
     val id: Int = 0,
 
@@ -11,7 +11,7 @@ data class Challenge(
 
     val isComplete: Boolean = false,
 
-    val minutes: Int = 0,
+    val minutes: Double = 0.0,
 
     val reward: String = "",
